Remove nodemon as a dependency
Getting nodemon + ngrok to work in concert with one another was painful and likely not worth the time, so it's being removed as the process runner.

This means that every time a change occurs the server must be restarted, but when the server is started it will always configure the webhook endpoint correctly.